Add files via upload
This feature is currently designed to facilitate forced patching of 1.20.6 files, ensuring server consistency. However, it is intended to serve a broader purpose in the future by enabling compatibility with other server versions. The envisioned expansion includes adding support for specific mobs and items tailored for administrators using different server versions. This commit lays the groundwork for future enhancements to ensure backward and forward compatibility with upcoming server versions.